requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.43.0
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
opencv-python==4.8.1.78
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0

tensorflow~=2.13.0
keras~=2.13.0
scikit-learn~=1.5.0
seaborn~=0.12.2

fastapi
uvicorn
joblib~=1.4.2

python-multipart
PyYAML~=6.0.2